analysis/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.0.8
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0.post0
langdetect==1.0.9
rapidfuzz==3.9.7
vaderSentiment==3.3.2
scikit-learn==1.5.2
spacy==3.7.6
python-multipart==0.0.9
numpy==1.26.4
pytest==8.3.3
